+++ clickhouse-data/src/test/java/com/clickhouse/data/ClickHouseUtilsTest.java
@@ -3,6 +3,7 @@ package com.clickhouse.data;
 import java.io.File;
 import java.io.FileNotFoundException;
 import java.io.IOException;
+import java.nio.file.Paths;
 import java.util.Arrays;
 import java.util.Collections;
 import java.util.HashMap;
@@ -55,6 +56,34 @@ public class ClickHouseUtilsTest {
         Assert.assertEquals(ClickHouseUtils.extractParameters("*&a=1&!b", null), expected);
         Assert.assertEquals(ClickHouseUtils.extractParameters("*&a=1&!b", new HashMap<>()), expected);
         Assert.assertEquals(ClickHouseUtils.extractParameters("*&a=1&!b", expected), expected);
+    }
+
+    @Test(groups = { "unit" })
+    public void testGetFile() throws IOException {
+        Assert.assertThrows(IllegalArgumentException.class, () -> ClickHouseUtils.getFile(null));
+        Assert.assertThrows(IllegalArgumentException.class, () -> ClickHouseUtils.getFile(""));
+
+        Assert.assertTrue(ClickHouseUtils.getFile("README.md").toFile().exists());
+
+        Assert.assertTrue(ClickHouseUtils.getFile("README.md").toFile().exists());
+        Assert.assertTrue(ClickHouseUtils.getFile("../README.md").toFile().exists());
+
+        Assert.assertEquals(ClickHouseUtils.getFile("~/a.csv"), Paths.get(System.getProperty("user.home"), "a.csv"));
+        Assert.assertEquals(ClickHouseUtils.getFile("~/../a.csv"),
+                Paths.get(System.getProperty("user.home"), "../a.csv").normalize());
+    }
+
+    @Test(groups = { "unit" })
+    public void testFindFiles() throws IOException {
+        Assert.assertThrows(IllegalArgumentException.class, () -> ClickHouseUtils.findFiles(null));
+        Assert.assertThrows(IllegalArgumentException.class, () -> ClickHouseUtils.findFiles(""));
+
+        Assert.assertEquals(ClickHouseUtils.findFiles("README.md").size(), 1);
+        Assert.assertEquals(ClickHouseUtils.findFiles("READM?.md").size(), 1);
+        Assert.assertEquals(ClickHouseUtils.findFiles("glob:*.md").size(), 1);
+        Assert.assertTrue(ClickHouseUtils.findFiles("glob:**.java", "src", "..").size() >= 1);
+        Assert.assertTrue(ClickHouseUtils.findFiles("glob:**.java", "src/test").size() >= 1);
+        Assert.assertTrue(ClickHouseUtils.findFiles("glob:../*.md", "../").size() >= 1);
     }
 
     @Test(groups = { "unit" })