+++ clickhouse-jdbc/src/test/java/com/clickhouse/jdbc/JdbcParseHandlerTest.java
@@ -13,6 +13,51 @@ import org.testng.annotations.Test;
 
 public class JdbcParseHandlerTest {
     @Test(groups = "unit")
+    public void testInsertFromInFileStatement() {
+        JdbcParseHandler handler = JdbcParseHandler.getInstance(false, false, true);
+        Assert.assertEquals(ClickHouseSqlParser.parse("INSERT INTO aaa", null, handler)[0].getSQL(), "INSERT INTO aaa");
+        Assert.assertEquals(ClickHouseSqlParser.parse("INSERT INTO aaa INFILE", null, handler)[0].getSQL(),
+                "INSERT INTO aaa INFILE"); // invalid
+        Assert.assertEquals(ClickHouseSqlParser.parse("INSERT INTO aaa FROM INFILE", null, handler)[0].getSQL(),
+                "INSERT INTO aaa FROM INFILE"); // invalid
+        Assert.assertEquals(ClickHouseSqlParser.parse("INSERT INTO aaa FROM INFILE 'a.csv'", null, handler)[0].getSQL(),
+                "INSERT INTO aaa FORMAT CSV");
+        Assert.assertEquals(
+                ClickHouseSqlParser.parse("INSERT INTO aaa FROM INFILE 'a.csv' Format CSV", null, handler)[0].getSQL(),
+                "INSERT INTO aaa Format CSV");
+        Assert.assertEquals(
+                ClickHouseSqlParser.parse("INSERT INTO aaa FROM INFILE 'a.csv' settings a=2", null, handler)[0]
+                        .getSQL(),
+                "INSERT INTO aaa settings a=2");
+        Assert.assertEquals(
+                ClickHouseSqlParser.parse("INSERT INTO aaa FROM INFILE 'a.csv.gz' compression 'gzip' settings a=2",
+                        null, handler)[0]
+                        .getSQL(),
+                "INSERT INTO aaa settings a=2");
+        Assert.assertEquals(
+                ClickHouseSqlParser.parse(
+                        "INSERT INTO aaa FROM INFILE 'input_*.csv.gz' compression 'gzip' settings max_result_rows=1, max_execution_time=2 FORMAT CSV",
+                        null, handler)[0].getSQL(),
+                "INSERT INTO aaa settings max_result_rows=1, max_execution_time=2 FORMAT CSV");
+    }
+
+    @Test(groups = "unit")
+    public void testSelectIntoOutFileStatement() {
+        JdbcParseHandler handler = JdbcParseHandler.getInstance(false, false, true);
+        Assert.assertEquals(ClickHouseSqlParser.parse("select 1", null, handler)[0].getSQL(), "select 1");
+        Assert.assertEquals(ClickHouseSqlParser.parse("select * from outfile", null, handler)[0].getSQL(),
+                "select * from outfile");
+        Assert.assertEquals(ClickHouseSqlParser.parse("select into outfile", null, handler)[0].getSQL(),
+                "select into outfile"); // invalid
+        Assert.assertEquals(ClickHouseSqlParser.parse("select 1 into outfile 'a.csv'", null, handler)[0].getSQL(),
+                "select 1 ");
+        Assert.assertEquals(ClickHouseSqlParser.parse(
+                "select * from numbers(10) settings max_result_rows=1,max_execution_time=3 into outfile 'a.csv' format CSV",
+                null, handler)[0].getSQL(),
+                "select * from numbers(10) settings max_result_rows=1,max_execution_time=3 format CSV");
+    }
+
+    @Test(groups = "unit")
     public void testParseDeleteStatement() {
         Assert.assertEquals(
                 ClickHouseSqlParser.parse("delete from tbl", null, JdbcParseHandler.INSTANCE)[0]
@@ -47,7 +92,7 @@ public class JdbcParseHandlerTest {
         Assert.assertEquals(
                 ClickHouseSqlParser.parse("update tbl set a=1,b=2 where 1", null,
                         JdbcParseHandler.INSTANCE)[0]
-                                .getSQL(),
+                        .getSQL(),
                 "ALTER TABLE `tbl` UPDATE a=1,b=2 where 1 SETTINGS mutations_sync=1");
         Assert.assertEquals(
                 ClickHouseSqlParser.parse(
@@ -64,7 +109,7 @@ public class JdbcParseHandlerTest {
     @Test(groups = "unit")
     public void testDeleteStatementWithoutWhereClause() {
         Assert.assertEquals(JdbcParseHandler.INSTANCE.handleStatement("delete  from  `a\\`' a` . tbl",
-                StatementType.DELETE, null, "a\\`' a", "tbl", null, null, null, null,
+                StatementType.DELETE, null, "a\\`' a", "tbl", null, null, null, null, null, null,
                 new HashMap<String, Integer>() {
                     {
                         put("DELETE", 0);
@@ -72,19 +117,17 @@ public class JdbcParseHandlerTest {
                     }
                 }, null, null),
                 new ClickHouseSqlStatement("TRUNCATE TABLE  `a\\`' a` . tbl", StatementType.DELETE,
-                        null,
-                        "a\\`' a", "tbl", null, null, null, null, null, null, null));
+                        null, "a\\`' a", "tbl", null, null, null, null, null, null, null, null, null));
         Assert.assertEquals(JdbcParseHandler.INSTANCE.handleStatement("delete from  `table\\`'1`",
                 StatementType.DELETE,
-                null, null, "table1", null, null, null, null, new HashMap<String, Integer>() {
+                null, null, "table1", null, null, null, null, null, null, new HashMap<String, Integer>() {
                     {
                         put("DELETE", 0);
                         put("FROM", 7);
                     }
                 }, null, null),
                 new ClickHouseSqlStatement("TRUNCATE TABLE  `table\\`'1`", StatementType.DELETE, null,
-                        null,
-                        "table1", null, null, null, null, null, null, null),
+                        null, "table1", null, null, null, null, null, null, null, null, null),
                 null);
     }
 
@@ -96,23 +139,23 @@ public class JdbcParseHandlerTest {
         positions.put("WHERE", 28);
         Assert.assertEquals(
                 JdbcParseHandler.INSTANCE.handleStatement("delete from  `a\\`' a` . tbl where a = b",
-                        StatementType.DELETE, null, "a\\`' a", "tbl", null, null, null, null,
-                        positions, null, null),
+                        StatementType.DELETE, null, "a\\`' a", "tbl", null, null, null, null, null, null, positions,
+                        null, null),
                 new ClickHouseSqlStatement(
                         "ALTER TABLE `a\\`' a`.`tbl` DELETE where a = b SETTINGS mutations_sync=1",
-                        StatementType.DELETE, null, "a\\`' a", "tbl", null, null, null, null,
-                        null, null, null));
+                        StatementType.DELETE, null, "a\\`' a", "tbl", null, null, null, null, null, null, null, null,
+                        null));
         positions.put("DELETE", 0);
         positions.put("FROM", 8);
         positions.put("WHERE", 26);
         Assert.assertEquals(
                 JdbcParseHandler.INSTANCE.handleStatement("delete  from  `table\\`'1` where 1",
-                        StatementType.DELETE,
-                        null, null, "table\\`'1", null, null, null, null, positions, null, null),
+                        StatementType.DELETE, null, null, "table\\`'1", null, null, null, null, null, null, positions,
+                        null, null),
                 new ClickHouseSqlStatement(
                         "ALTER TABLE `table\\`'1` DELETE where 1 SETTINGS mutations_sync=1",
-                        StatementType.DELETE, null, null, "table\\`'1", null, null, null, null,
-                        null, null, null));
+                        StatementType.DELETE, null, null, "table\\`'1", null, null, null, null, null,
+                        null, null, null, null));
     }
 
     @Test(groups = "unit")
@@ -128,11 +171,9 @@ public class JdbcParseHandlerTest {
         Map<String, String> settings = Collections.singletonMap("a", "1");
         Assert.assertEquals(
                 JdbcParseHandler.INSTANCE.handleStatement(sql1, StatementType.DELETE, null, null, "tbl",
-                        null, null,
-                        null, null, positions, settings, null),
+                        null, null, null, null, null, null, positions, settings, null),
                 new ClickHouseSqlStatement("TRUNCATE TABLE tbl settings a=1", StatementType.DELETE,
-                        null, null, "tbl",
-                        null, null, null, null, null, settings, null));
+                        null, null, "tbl", null, null, null, null, null, null, null, settings, null));
 
         String sql2 = "delete from tbl where a != 1 and b != 2 settings a=1,b='a'";
         positions = new HashMap<String, Integer>() {
@@ -151,11 +192,10 @@ public class JdbcParseHandlerTest {
         };
         Assert.assertEquals(
                 JdbcParseHandler.INSTANCE.handleStatement(sql2, StatementType.DELETE, null, null, "tbl",
-                        null, null,
-                        null, null, positions, settings, null),
+                        null, null, null, null, null, null, positions, settings, null),
                 new ClickHouseSqlStatement(
                         "ALTER TABLE `tbl` DELETE where a != 1 and b != 2 SETTINGS mutations_sync=1, a=1,b='a'",
-                        StatementType.DELETE, null, null, "tbl", null, null, null, null, null,
+                        StatementType.DELETE, null, null, "tbl", null, null, null, null, null, null, null,
                         settings, null));
 
         String sql3 = "delete from tbl where a != 1 and b != 2 settings a=1,mutations_sync=2,b='a'";
@@ -176,18 +216,17 @@ public class JdbcParseHandlerTest {
         };
         Assert.assertEquals(
                 JdbcParseHandler.INSTANCE.handleStatement(sql3, StatementType.DELETE, null, null, "tbl",
-                        null, null,
-                        null, null, positions, settings, null),
+                        null, null, null, null, null, null, positions, settings, null),
                 new ClickHouseSqlStatement(
                         "ALTER TABLE `tbl` DELETE where a != 1 and b != 2 settings a=1,mutations_sync=2,b='a'",
-                        StatementType.DELETE, null, null, "tbl", null, null, null, null, null,
-                        settings, null));
+                        StatementType.DELETE, null, null, "tbl", null, null, null, null, null, null, null, settings,
+                        null));
     }
 
     @Test(groups = "unit")
     public void testUpdateStatementWithoutWhereClause() {
         Assert.assertEquals(JdbcParseHandler.INSTANCE.handleStatement("update  `a\\`' a` . tbl set a=1",
-                StatementType.UPDATE, null, "a\\`' a", "tbl", null, null, null, null,
+                StatementType.UPDATE, null, "a\\`' a", "tbl", null, null, null, null, null, null,
                 new HashMap<String, Integer>() {
                     {
                         put("UPDATE", 0);
@@ -196,10 +235,10 @@ public class JdbcParseHandlerTest {
                 }, null, null),
                 new ClickHouseSqlStatement(
                         "ALTER TABLE `a\\`' a`.`tbl` UPDATE a=1 SETTINGS mutations_sync=1",
-                        StatementType.UPDATE, null, "a\\`' a", "tbl", null, null, null, null,
-                        null, null, null));
+                        StatementType.UPDATE, null, "a\\`' a", "tbl", null, null, null, null, null, null, null, null,
+                        null));
         Assert.assertEquals(JdbcParseHandler.INSTANCE.handleStatement("update  `table\\`'1` set a=1",
-                StatementType.UPDATE, null, null, "table1", null, null, null, null,
+                StatementType.UPDATE, null, null, "table1", null, null, null, null, null, null,
                 new HashMap<String, Integer>() {
                     {
                         put("UPDATE", 0);
@@ -207,8 +246,8 @@ public class JdbcParseHandlerTest {
                     }
                 }, null, null),
                 new ClickHouseSqlStatement("ALTER TABLE `table1` UPDATE a=1 SETTINGS mutations_sync=1",
-                        StatementType.UPDATE, null, null, "table1", null, null, null, null,
-                        null, null, null));
+                        StatementType.UPDATE, null, null, "table1", null, null, null, null, null, null, null, null,
+                        null));
     }
 
     @Test(groups = "unit")
@@ -219,22 +258,22 @@ public class JdbcParseHandlerTest {
         Assert.assertEquals(
                 JdbcParseHandler.INSTANCE.handleStatement(
                         "Update  `a\\`' a` . tbl set a = 2 where a = b",
-                        StatementType.UPDATE, null, "a\\`' a", "tbl", null, null, null, null,
-                        positions, null, null),
+                        StatementType.UPDATE, null, "a\\`' a", "tbl", null, null, null, null, null, null, positions,
+                        null, null),
                 new ClickHouseSqlStatement(
                         "ALTER TABLE `a\\`' a`.`tbl` UPDATE a = 2 where a = b SETTINGS mutations_sync=1",
-                        StatementType.UPDATE, null, "a\\`' a", "tbl", null, null, null, null,
-                        null, null, null));
+                        StatementType.UPDATE, null, "a\\`' a", "tbl", null, null, null, null, null, null, null, null,
+                        null));
         positions.put("UPDATE", 0);
         positions.put("SET", 19);
         Assert.assertEquals(
                 JdbcParseHandler.INSTANCE.handleStatement("update `table\\`'1` set a = b where 1",
-                        StatementType.UPDATE,
-                        null, null, "table\\`'1", null, null, null, null, positions, null, null),
+                        StatementType.UPDATE, null, null, "table\\`'1", null, null, null, null, null, null, positions,
+                        null, null),
                 new ClickHouseSqlStatement(
                         "ALTER TABLE `table\\`'1` UPDATE a = b where 1 SETTINGS mutations_sync=1",
-                        StatementType.UPDATE, null, null, "table\\`'1", null, null, null, null,
-                        null, null, null));
+                        StatementType.UPDATE, null, null, "table\\`'1", null, null, null, null, null, null, null, null,
+                        null));
     }
 
     @Test(groups = "unit")
@@ -250,12 +289,11 @@ public class JdbcParseHandlerTest {
         Map<String, String> settings = Collections.singletonMap("a", "1");
         Assert.assertEquals(
                 JdbcParseHandler.INSTANCE.handleStatement(sql1, StatementType.UPDATE, null, null, "tbl",
-                        null, null,
-                        null, null, positions, settings, null),
+                        null, null, null, null, null, null, positions, settings, null),
                 new ClickHouseSqlStatement(
                         "ALTER TABLE `tbl` UPDATE x=1 SETTINGS mutations_sync=1, a=1",
-                        StatementType.UPDATE, null, null, "tbl", null, null, null, null, null,
-                        settings, null));
+                        StatementType.UPDATE, null, null, "tbl", null, null, null, null, null, null, null, settings,
+                        null));
 
         String sql2 = "update tbl set x=1, y=2 where a != 1 and b != 2 settings a=1,b='a'";
         positions = new HashMap<String, Integer>() {
@@ -274,12 +312,11 @@ public class JdbcParseHandlerTest {
         };
         Assert.assertEquals(
                 JdbcParseHandler.INSTANCE.handleStatement(sql2, StatementType.UPDATE, null, null, "tbl",
-                        null, null,
-                        null, null, positions, settings, null),
+                        null, null, null, null, null, null, positions, settings, null),
                 new ClickHouseSqlStatement(
                         "ALTER TABLE `tbl` UPDATE x=1, y=2 where a != 1 and b != 2 SETTINGS mutations_sync=1, a=1,b='a'",
-                        StatementType.UPDATE, null, null, "tbl", null, null, null, null, null,
-                        settings, null));
+                        StatementType.UPDATE, null, null, "tbl", null, null, null, null, null, null, null, settings,
+                        null));
 
         String sql3 = "update tbl set x=1,y=2 where a != 1 and b != 2 settings a=1,mutations_sync=2,b='a'";
         positions = new HashMap<String, Integer>() {
@@ -299,11 +336,10 @@ public class JdbcParseHandlerTest {
         };
         Assert.assertEquals(
                 JdbcParseHandler.INSTANCE.handleStatement(sql3, StatementType.UPDATE, null, null, "tbl",
-                        null, null,
-                        null, null, positions, settings, null),
+                        null, null, null, null, null, null, positions, settings, null),
                 new ClickHouseSqlStatement(
                         "ALTER TABLE `tbl` UPDATE x=1,y=2 where a != 1 and b != 2 settings a=1,mutations_sync=2,b='a'",
-                        StatementType.UPDATE, null, null, "tbl", null, null, null, null, null,
-                        settings, null));
+                        StatementType.UPDATE, null, null, "tbl", null, null, null, null, null, null, null, settings,
+                        null));
     }
 }